+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 - fix: prevent crashes from non JSON-serializable property types (Date, URL, Data, infinity, NaN, etc.) ([#466](https://github.com/PostHog/posthog-ios/pull/466))
+- feat: add `$is_testflight` and `$is_sideloaded` event properties ([#443](https://github.com/PostHog/posthog-ios/pull/443))
 
 ## 3.40.0 - 2026-02-05